Struct fringe::Context

pub struct Context<Stack: Stack> {
    // some fields omitted
}

Context holds a suspended thread of execution along with a stack.

It can be swapped into and out of with the swap method, and once you're done with it, you can get the stack back through unwrap.

Every operation is unsafe, because no guarantees can be made about the state of the context.

Methods

impl<Stack> Context<Stack> where Stack: Stack

unsafe fn new(stack: Stack, f: unsafe extern fn(usize) -> !) -> Context<Stack>

Creates a new Context. When it is swapped into, it will call f(arg), where arg is the argument passed to swap.

unsafe fn unwrap(self) -> Stack

Unwraps the context, returning the stack it contained.

impl<OldStack> Context<OldStack> where OldStack: Stack

unsafe fn swap<NewStack>(old_ctx: *mut Context<OldStack>, new_ctx: *const Context<NewStack>, arg: usize) -> usize where NewStack: Stack

Switches to in_ctx, saving the current thread of execution to out_ctx.