Struct fringe::OsStack

pub struct OsStack { /* fields omitted */ }

OsStack holds a guarded stack allocated using the operating system's anonymous memory mapping facility.

Methods

impl OsStack

fn new(size: usize) -> Result<OsStack, IoError>

Allocates a new stack with at least size accessible bytes. size is rounded up to an integral number of pages; OsStack::new(0) is legal and allocates the smallest possible stack, consisting of one data page and one guard page.