Struct fringe::SliceStack

pub struct SliceStack<'a>(_);

SliceStack holds a non-guarded stack allocated elsewhere and provided as a mutable slice.

Methods

impl<'a> SliceStack<'a>

fn new(slice: &'a mut [u8]) -> SliceStack<'a>

Creates a SliceStack from an existing slice.

This function will automatically align the slice to make it suitable for use as a stack. However this function may panic if the slice is smaller than STACK_ALIGNMENT.