Trait fringe::GuardedStack

pub unsafe trait GuardedStack { }

A marker trait for Stack objects with a guard page.

To preserve memory safety, an implementation of this trait must fulfill the following contract, in addition to the contract of Stack:

Any access of data at addresses limit() to limit().offset(4096) must abnormally terminate, at least, the thread that performs the access.