Struct fringe::OsStack

pub struct OsStack {
    // some fields omitted
}

This object represents a stack from the operating system's anonymous memory mapping facility, usually mmap. The stack it provides comes with a guard page, which is not included in the stack limit.

Methods

impl Stack

fn new(size: usize) -> Result<Stack, IoError>